requirements.txt
flask==3.0.2
flask-limiter==3.5.0
python-dotenv>=1.0.0
openai==1.12.0
prometheus-client==0.20.0
aiohttp>=3.9.0
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
google-search-results==2.4.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
tenacity==8.2.3
asyncio>=3.4.3
certifi>=2024.2.2 